Move load_params() to snek5000.params module
